requirements.txt
Flask==3.1.0
google-auth==2.38.0
google-cloud-aiplatform==1.79.0
google-api-python-client==2.159.0
python-dotenv==1.0.1
slack_sdk==3.34.0
librosa==0.10.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.4
openai-whisper==20240930
pydub==0.25.1
requests==2.32.3
python-dateutil~=2.9.0.post0
numpy~=2.1.3
pysbd~=0.3.4
scikit-learn~=1.6.1
scipy~=1.15.1